chatbot/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
groq==0.13.1
supabase==2.10.0
vaderSentiment==3.3.2
python-dotenv==1.0.1
